+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- PartMaterialIngredient can now exclude specific material categories (JaisDK) [#780]
+### Changed
+- Updated Japanese translation (twister716) [#750]
+### Fixed
+- Allow salvager to keep running even when output isn't empty (kylev) [#768]
+
 ## [1.21.1-4.0.12] - 2025-01-25
 ### Fixed
 - Part names not translating in some cases (kylev) [#767]